add ser/de

+++ core/src/error.rs
@@ -3,7 +3,7 @@ use thiserror::Error;
 #[derive(Debug, Error)]
 pub enum Error {
     #[error("wrong bytes index {0}")]
-    WrongBytesIndex(u8),
+    WrongBytesIndex(usize),
 
     #[error("wrong type {0} {1}")]
     WrongType(&'static str, u8),
